cleanup + added ManyToMany field to SigningKey model
