+++ zapisy/apps/grade/ticket_create/management/commands/create_graded.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 from apps.grade.ticket_create.models.student_graded import StudentGraded
-from apps.grade.ticket_create.models.used_ticket_stamp import UsedTicketStamp
+from apps.grade.ticket_create.models.signing_key import SigningKey
 
 
 class Command(BaseCommand):
@@ -8,5 +8,6 @@ class Command(BaseCommand):
     help = 'tworzy ankiety na podstawie szablony'
 
     def handle(self, *args, **options):
-        for p in UsedTicketStamp.objects.select_related('student', 'poll', 'poll__semester').all():
-            st, c = StudentGraded.objects.get_or_create(student=p.student, semester=p.poll.semester)
+        for signingkey in SigningKey.objects.select_related('students', 'poll', 'poll__semester').all():
+            for student in signingkey.students.objects.all():
+            st, c = StudentGraded.objects.get_or_create(student=student, semester=signingkey.poll.semester)
