+++ zapisy/apps/grade/ticket_create/models/__init__.py
@@ -1,4 +1,2 @@
-from .used_ticket_stamp import UsedTicketStamp
-
 from .signing_key import SigningKey
 from .student_graded import StudentGraded
