--- zapisy/apps/grade/ticket_create/models/used_ticket_stamp.py
@@ -1,24 +1,0 @@
-from django.db import models
-from apps.users.models import Student
-
-
-class UsedTicketStamp(models.Model):
-    from apps.grade.poll.models.poll import Poll
-    student = models.ForeignKey(Student, on_delete=models.CASCADE, verbose_name="student")
-    poll = models.ForeignKey(Poll, on_delete=models.CASCADE, verbose_name="ankieta")
-
-    class Meta:
-        verbose_name = 'wykorzystany bilet'
-        verbose_name_plural = 'wykorzystane bilety'
-        app_label = 'ticket_create'
-
-    def __str__(self):
-        return str(self.student) + " " + str(self.poll)
-
-    @staticmethod
-    def check_exists(student, poll):
-        return len(UsedTicketStamp.objects.filter(student=student, poll=poll)) == 0
-
-    @staticmethod
-    def get_grading_students_ids():
-        return UsedTicketStamp.objects.all().values_list('student', flat=True).distinct()
