+++ zapisy/apps/grade/ticket_create/serializers.py
@@ -5,5 +5,6 @@ class SigningRequestSerializer(serializers.Serializer):
     id = serializers.IntegerField(required=True, min_value=0)
     ticket = serializers.IntegerField(required=True, min_value=0)
 
+
 class SigningRequestsListSerializer(serializers.Serializer):
     signing_requests = serializers.ListField(child=SigningRequestSerializer())
